Implement very rough MVP action checks for the entire unit page

+++ service/src/main/kotlin/fi/espoo/evaka/daycare/controllers/DaycareController.kt
@@ -7,7 +7,6 @@ package fi.espoo.evaka.daycare.controllers
 import fi.espoo.evaka.Audit
 import fi.espoo.evaka.daycare.Daycare
 import fi.espoo.evaka.daycare.DaycareFields
-import fi.espoo.evaka.daycare.DaycareGroupSummary
 import fi.espoo.evaka.daycare.controllers.utils.created
 import fi.espoo.evaka.daycare.controllers.utils.noContent
 import fi.espoo.evaka.daycare.controllers.utils.ok
@@ -75,7 +74,15 @@ class DaycareController(
         return db.read { tx ->
             tx.getDaycare(daycareId)?.let { daycare ->
                 val groups = tx.getDaycareGroupSummaries(daycareId)
-                ResponseEntity.ok(DaycareResponse(daycare, groups, accessControl.getPermittedUnitActions(user, daycareId)))
+                ResponseEntity.ok(
+                    DaycareResponse(
+                        daycare,
+                        groups.map {
+                            DaycareGroupResponse(id = it.id, name = it.name, permittedActions = accessControl.getPermittedGroupActions(user, it.id))
+                        },
+                        accessControl.getPermittedUnitActions(user, daycareId)
+                    )
+                )
             }
         } ?: ResponseEntity.notFound().build()
     }
@@ -109,8 +116,7 @@ class DaycareController(
         @RequestBody body: CreateGroupRequest
     ): ResponseEntity<DaycareGroup> {
         Audit.UnitGroupsCreate.log(targetId = daycareId)
-        acl.getRolesForUnit(user, daycareId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Unit.CREATE_GROUP, daycareId)
 
         return db.transaction { daycareService.createGroup(it, daycareId, body.name, body.startDate, body.initialCaretakers) }
             .let { created(it, URI.create("/$daycareId/groups/${it.id}")) }
@@ -130,8 +136,7 @@ class DaycareController(
         @RequestBody body: GroupUpdateRequest
     ): ResponseEntity<Unit> {
         Audit.UnitGroupsUpdate.log(targetId = groupId)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Group.UPDATE, groupId)
 
         db.transaction { it.updateGroup(groupId, body.name, body.startDate, body.endDate) }
 
@@ -146,8 +151,7 @@ class DaycareController(
         @PathVariable("groupId") groupId: GroupId
     ): ResponseEntity<Unit> {
         Audit.UnitGroupsDelete.log(targetId = groupId)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Group.DELETE, groupId)
 
         db.transaction { daycareService.deleteGroup(it, groupId) }
         return noContent()
@@ -161,8 +165,7 @@ class DaycareController(
         @PathVariable("groupId") groupId: GroupId
     ): ResponseEntity<CaretakersResponse> {
         Audit.UnitGroupsCaretakersRead.log(targetId = groupId)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF, UserRole.SPECIAL_EDUCATION_TEACHER)
+        accessControl.requirePermissionFor(user, Action.Group.READ_CARETAKERS, groupId)
 
         return db.read {
             val daycareStub = it.getDaycareStub(daycareId)
@@ -185,8 +188,7 @@ class DaycareController(
         @RequestBody body: CaretakerRequest
     ): ResponseEntity<Unit> {
         Audit.UnitGroupsCaretakersCreate.log(targetId = groupId)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Group.CREATE_CARETAKERS, groupId)
 
         db.transaction {
             caretakerService.insert(
@@ -210,8 +212,7 @@ class DaycareController(
         @RequestBody body: CaretakerRequest
     ): ResponseEntity<Unit> {
         Audit.UnitGroupsCaretakersUpdate.log(targetId = id)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Group.UPDATE_CARETAKERS, groupId)
 
         db.transaction {
             caretakerService.update(
@@ -235,8 +236,7 @@ class DaycareController(
         @PathVariable("id") id: UUID
     ): ResponseEntity<Unit> {
         Audit.UnitGroupsCaretakersDelete.log(targetId = id)
-        acl.getRolesForUnitGroup(user, groupId)
-            .requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR)
+        accessControl.requirePermissionFor(user, Action.Group.DELETE_CARETAKERS, groupId)
 
         db.transaction {
             caretakerService.delete(
@@ -325,6 +325,7 @@ class DaycareController(
         val groupName: String,
         val caretakers: List<CaretakerAmount>
     )
-}
 
-data class DaycareResponse(val daycare: Daycare, val groups: List<DaycareGroupSummary>, val permittedActions: Set<Action.Unit>)
+    data class DaycareGroupResponse(val id: GroupId, val name: String, val permittedActions: Set<Action.Group>)
+    data class DaycareResponse(val daycare: Daycare, val groups: List<DaycareGroupResponse>, val permittedActions: Set<Action.Unit>)
+}
